Updated the project to use the most current SDK.  Set the target OS to 10.9.  Replaced depreciated methods.  Converted to ARC and turned that on.  (Except for two source code files)  Imported in the LOST AJR Frameworks classes for AJRClassEnumerator and AJRObjectBroker.  I previoiusly guessed what AJRObjectBroker did, and was pretty far off.  Some things now work where they did not before.  And also some things DON'T work where they did before.  This is the first step to getting this application working better than it currently is.

+++ EOModeler/Document-OutlineView.m
@@ -10,7 +10,7 @@
 
 @implementation Document (OutlineView)
 
-- (int)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
+- (NSInteger)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
 {
 	if (item == nil) {
 		return 1;
@@ -53,7 +53,7 @@
 	return NO;
 }
 
-- (id)outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item
+- (id)outlineView:(NSOutlineView *)outlineView child:(NSInteger)index ofItem:(id)item
 {
 	if (item == nil) {
 		return model;
